selftests: forwarding, mlxsw: Use slave MAC for bridge address
When the first netdevice is enslaved to a bridge, the bridge inherits MAC
address from the slave. However, this only happens after the enslavement is
acked by the drivers.

In the future, mlxsw will start creating a router interface (RIF) for
netdevices that become mlxsw uppers, if they are L3 netdevices (i.e. have
an IP address). Previously, this did not happen. At the point when the
enslavement is evaluated, the bridge MAC address is still whatever the
bridge defaults to. Thus it seems as if that would be the MAC address of
the newly-created RIF. But all RIFs must share one of N common prefixes,
where N is ASIC-dependent. For N of one, the prefix of the default bridge
address very likely does not agree with the prefix of the other existing
RIFs. The new RIF thus cannot be created, which means the configuration is
not offloadable, and the proposal to enslave the port to the bridge is
vetoed.

Working around this in code is probably not reasonable, because it would
need to make assumptions about what the MAC address will change to. A
number of (mainly VXLAN) selftests already set a bridge address to a front
panel port netdevice address. Do the same for two more tests.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Reviewed-by: Amit Cohen <amcohen@nvidia.com>