selftests: mlxsw: q_in_q_veto: Disable IPv6 autogen on bridges
In a future patch, mlxsw will start adding RIFs to uppers of front panel
port netdevices, if they have an IP address. This will cause the veto
selftests to fail in a wrong manner. The swp enslavement to the 802.1ad
bridge is not allowed, because RIFs are not allowed to be created for
802.1ad bridges, but the address indicates one needs to be created.
Then the attempt to create a VLAN on top of the same bridge is not vetoed,
because the bridge is not related to mlxsw, and the selftest fails.

Fix by disabling automatic IPv6 address generation for the bridges in this
selftest, thus exempting them from the mlxsw router attention.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Reviewed-by: Amit Cohen <amcohen@nvidia.com>