selftests: forwarding: q_in_vni: Disable IPv6 autogen on bridges
In a future patch, mlxsw will start adding RIFs to uppers of front panel
port netdevices, if they have an IP address. This will cause this selftest
to fail spuriously. The swp enslavement to the 802.1ad bridge is not
allowed, because RIFs are not allowed to be created for 802.1ad bridges,
but the address indicates one needs to be created.

Fix by disabling automatic IPv6 address generation for the HW-offloaded
bridge in this selftest, thus exempting it from mlxsw router attention.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Reviewed-by: Amit Cohen <amcohen@nvidia.com>