mlxsw: pci: Add support for new reset flow
The driver resets the device during probe and during a devlink reload.
The current reset method reloads the current firmware version or a pending
one, if one was previously flashed using devlink. However, the reset does
not take down the PCI link, preventing the PCI firmware from being
upgraded, unless the system is rebooted.

To solve this problem, a new reset command (6) was implemented in the
firmware. Unlike the current command (1), after issuing the new command
the device will not start the reset immediately, but only after the PCI
link was disabled. The driver is expected to wait for 500ms before
re-enabling the link to give the firmware enough time to start the reset.

Implement the new reset method and use it only after verifying it is
supported by the current firmware version by querying the Management
Capabilities Mask (MCAM) register. Consider the PCI firmware to be
operational either after waiting for a predefined time of 2000ms or after
reading an active link status when "Data Link Layer Link Active Reporting"
is supported. For good measures, make sure the device ID can be read from
the configuration space of the device.

Once the PCI firmware is operational, go back to the regular reset flow
and wait for the entire device to become ready. That is, repeatedly read
the "system_status" register from the BAR until a value of "FW_READY"
(0x5E) appears.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>
Signed-off-by: Petr Machata <petrm@nvidia.com>