PCI: Add no bus reset quirk for NVIDIA Spectrum-3 devices
On NVIDIA Spectrum-3 devices the device is reset following a bus reset,
but the firmware does not recover from it, making it impossible for the
host to communicate with it. Prevent bus reset for these devices.

Specifically, I have observed that when polling on the system status in
mlxsw_pci_sys_ready_wait() following the reset, a status of 0xffff is
read indefinitely.

Before:

 # cat /sys/bus/pci/devices/0000\:06\:00.0/reset_method
 bus

After:

 # cat /sys/bus/pci/devices/0000\:06\:00.0/reset_method
 cat: '/sys/bus/pci/devices/0000:06:00.0/reset_method': No such file or directory

Signed-off-by: Ido Schimmel <idosch@nvidia.com>