mlxsw: pci: Add support for new reset flow
The driver resets the device during probe and during a devlink reload.
The current reset method reloads the current firmware version or a
pending one, if one was previously flashed using devlink. However, the
reset does not take down the PCI link, preventing the PCI firmware from
being upgraded, unless the system is rebooted.

To solve this problem, a new reset command (6) was implemented in the
firmware. Unlike the current command (1), after issuing the new command
the device will not start the reset immediately, but only after the PCI
link was disabled.

Implement the new reset method by first verifying that it is supported
by the current firmware version by querying the Management Capabilities
Mask (MCAM) register. If supported, issue the new reset command (6) via
MRSR register followed by a PCI reset. Specifically, trigger the PCI
reset by calling __pci_reset_function_locked() which will issue a bus
reset as it is the only reset method supported by the device.

Once the PCI firmware is operational, go back to the regular reset flow
and wait for the entire device to become ready. That is, repeatedly read
the "system_status" register from the BAR until a value of "FW_READY"
(0x5E) appears.

Tested:

 # for i in $(seq 1 10); do devlink dev reload pci/0000:01:00.0; done

 # devlink dev info pci/0000:01:00.0 | grep "fw.version"
         fw.version 34.2023.512

 # devlink dev flash pci/0000:01:00.0 file mellanox/fw-SPC4-rel-34_2022_0626-EVB.mfa2 &> /dev/null
 # devlink dev reload pci/0000:01:00.0
 # devlink dev info pci/0000:01:00.0 | grep "fw.version"
         fw.version 34.2022.626

 # devlink dev flash pci/0000:01:00.0 file mellanox/fw-SPC4-rel-34_2023_0512-EVB.mfa2 &> /dev/null
 # devlink dev reload pci/0000:01:00.0
 # devlink dev info pci/0000:01:00.0 | grep "fw.version"
         fw.version 34.2023.512

Signed-off-by: Ido Schimmel <idosch@nvidia.com>